Moved MapZoomButtons styled-components to MapZoomButtons/styles.js
