+++ src/components/MapZoomButtons/__tests__/ZoomButtonsStyledComponents.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { shallow } from "enzyme";
-import { ZoomButton, ZoomButtons } from "../ZoomButtons";
+import { ZoomButton, ZoomButtons } from "../styles";
 
 describe("tests for ZoomButtons", () => {
   test("render ZoomButton", () => {
